chore: refactor ship render
