+++ packages/client/src/layers/frontend/phaser/systems/index.ts
@@ -3,11 +3,11 @@ import { createActionSelectionSystem } from "./createActionSelectionSystem";
 import { createFireCannonAnimationSystem } from "./createFireCannonAnimationSystem";
 import { createInputSystem } from "./createInputSystem";
 import { createMoveOptionsSystem } from "./createMoveOptionsSystem";
-import { createPositionSystem } from "./createPositionSystem";
 import { createProjectionSystem } from "./createProjectionSystem";
 import { createRadiusSystem } from "./createRadiusSystem";
 import { createResetSystem } from "./createResetSystem";
 import { createShipCircleSystem } from "./createShipCircleSystem";
+import { createShipSystem } from "./createShipSystem";
 import { createStatAnimationSystem } from "./createStatAnimationSystem";
 import { createTargetedSystem } from "./createTargetedSystem";
 import { createTileSystem } from "./createTileSystem";
@@ -15,7 +15,7 @@ import { createTileSystem } from "./createTileSystem";
 export function createPhaserSystems(context: PhaserLayer) {
   createActionSelectionSystem(context);
   createInputSystem(context);
-  createPositionSystem(context);
+  createShipSystem(context);
   createProjectionSystem(context);
   createResetSystem(context);
   createRadiusSystem(context);
